Feat/category detection

+++ Newsportal/Controllers/NewsController.cs
@@ -11,6 +11,8 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.ML;
+using MlProject.DataModels;
 using Newsportal.ViewModels;
 
 namespace Newsportal.Controllers
@@ -21,13 +23,19 @@ namespace Newsportal.Controllers
         private readonly UserManager<IdentityUser> _userManager;
         private readonly IWebHostEnvironment webHostEnvironment;
         private readonly ILogger<NewsController> _logger;
+        private readonly PredictionEnginePool<NewsModel, CategoryPrediction> _predictionEnginePool;
 
-        public NewsController(ApplicationDbContext context, UserManager<IdentityUser> userManager, IWebHostEnvironment hostEnvironment, ILogger<NewsController> logger)
+        public NewsController(ApplicationDbContext context,
+            UserManager<IdentityUser> userManager,
+            IWebHostEnvironment hostEnvironment,
+            ILogger<NewsController> logger,
+            PredictionEnginePool<NewsModel,CategoryPrediction> predictionEnginePool)
         {
             _context = context;
             _userManager = userManager;
             webHostEnvironment = hostEnvironment;
             _logger = logger;
+            _predictionEnginePool = predictionEnginePool;
         }
 
         // GET: News
@@ -60,8 +68,12 @@ namespace Newsportal.Controllers
         // GET: News/Create
         public async Task<IActionResult> Create()
         {
-            ViewBag.Categories = new SelectList(await _context.Category.ToListAsync(), "Id", "CategoryName");
-            return View();
+            var newsCreateViewModel = new NewsCreateViewModel()
+            {
+                Categories = await _context.Category.ToListAsync()
+            };
+            //ViewBag.Categories = new SelectList(await _context.Category.ToListAsync(), "Id", "CategoryName");
+            return View(newsCreateViewModel);
         }
 
         // POST: News/Create
@@ -70,27 +82,61 @@ namespace Newsportal.Controllers
         [HttpPost]
         [ValidateAntiForgeryToken]
         [Authorize]
-        public async Task<IActionResult> Create([Bind("Id,Category,Title,Content,ImageFile,BreakingNews,FeaturedNews,IsPublished")] News news)
+        public async Task<IActionResult> Create([Bind("CategoryId,Title,Content,ImageFile,BreakingNews,FeaturedNews,IsPublished")] NewsCreateViewModel model)
         {
-            
-            //var errors = ModelState.Values.SelectMany(v => v.Errors).ElementAt(0);
-            news.Category = _context.Category.Where(a => a.Id == news.Category.Id).Single();
-
-            // errors = ModelState.Values.SelectMany(v => v.Errors).ElementAt(0);
-
-            
             if (!ModelState.IsValid)
             {
-                var user = (Reporter)await _userManager.GetUserAsync(this.User);
-                news.Reporter = user;
-                news.PublishedDate = DateTime.Now;
-                news.Image = UploadedFile(news);
+                return View(model);
+            }
+
+            try
+            {
+                var reporter = (Reporter)await _userManager.GetUserAsync(this.User);
+                Category category = null;
+            
+                //if Selected category is null, then predict the category of news
+                if (model.CategoryId == null)
+                {
+                    var input = new NewsModel()
+                    {
+                        Content = model.Content,
+                        Heading = model.Title
+                    };
+                    //here comes the prediction
+                    var prediction = _predictionEnginePool.Predict(modelName: "CategoryClassifier", example: input);
+                    var predictedCategory = prediction.PredictedLabel;
+                    category = await _context.Category.FirstOrDefaultAsync(c => c.CategoryName.ToLower() == predictedCategory.ToLower());
+                    if (category == null)
+                    {
+                        ModelState.AddModelError(string.Empty, "Category could not be predicted. Please enter category");
+                        model.Categories = await _context.Category.ToListAsync();
+                        return View(model);
+                    }
+
+                }
+                else
+                {
+                    category = await _context.Category.FindAsync(model.CategoryId);
+                }
+
+                var news = new News(DateTime.Now, reporter, category)
+                {
+                    Title = model.Title,
+                    Content = model.Content,
+                    Image = UploadedFile(model)
+
+                };
+
                 _context.News.Add(news);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
+
             }
-            ViewBag.Categories = new SelectList(await _context.Category.ToListAsync(), "Id", "CategoryName");
-            return View("create", news);
+            catch (DbUpdateException e)
+            {
+                _logger.LogError(e.Message);
+                return BadRequest("Something went wrong");
+            }
         }
 
         // GET: News/Edit/5
@@ -100,20 +146,35 @@ namespace Newsportal.Controllers
             {
                 return NotFound();
             }
-
-            ViewBag.Categories = new SelectList(await _context.Category.ToListAsync(), "Id", "CategoryName");
-
-
-            var news = await _context.News.FirstOrDefaultAsync(i => i.Id == id.Value);
-            var user = await _userManager.GetUserAsync(this.User);
-            /*if (news.IsPublished) { return Content("Cannot edit Published news"); }*/
-            //await db.Items.Include(i => i.ItemVerifications).FirstOrDefaultAsync(i => i.Id == id.Value);
-
+            
+            var news = await _context.News.FindAsync(id.Value);
+            var user = (Reporter) await _userManager.GetUserAsync(this.User);
             if (news == null || news.Reporter.Id != user.Id || news.LastEditedBy != null)
             {
                 return NotFound();
             }
-            return View(news);
+
+            if (!User.IsInRole("Admin") && news.IsPublished)
+            {
+                ModelState.AddModelError(string.Empty, "Cannot edit Published news");
+                return BadRequest("Cannot edit Published news");
+            }
+
+            var newsUpdateViewModel = new NewsUpdateViewModel()
+            {
+                Id = news.Id,
+                Image = news.Image,
+                PublishedDate = news.PublishedDate,
+                Categories = await _context.Category.ToListAsync(),
+                Content = news.Content,
+                Title = news.Title,
+                CategoryId = news.Category.Id,
+                IsBreaking = news.BreakingNews,
+                IsFeatured = news.FeaturedNews,
+                IsPublished = news.IsPublished
+            };
+
+            return View(newsUpdateViewModel);
         }
 
         // POST: News/Edit/5
@@ -121,35 +182,92 @@ namespace Newsportal.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Count,Category,PublishedDate,Title,Content,Image,BreakingNews,FeaturedNews,IsPublished")] News news)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,CategoryId,PublishedDate,Title,Content,ImageFile,IsBreaking,IsBreaking,IsPublished")] NewsUpdateViewModel model)
         {
-            if (id != news.Id)
+            if (id != model.Id)
             {
                 return NotFound();
             }
+
             if (!ModelState.IsValid)
             {
-                try
-                {
-                    news.Category = _context.Category.Where(a => a.Id == news.Category.Id).Single();
-
-                    _context.Update(news);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!NewsExists(news.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
+                return View(model);
             }
-            return View(news);
+
+            try
+            {
+                Category category = null;
+                if (model.CategoryId == null)
+                {
+                    var input = new NewsModel()
+                    {
+                        Content = model.Content,
+                        Heading = model.Title
+                    };
+                    //here comes the prediction
+                    var prediction = _predictionEnginePool.Predict(modelName: "CategoryClassifier", example: input);
+                    var predictedCategory = prediction.PredictedLabel;
+                    category = await _context.Category.FirstOrDefaultAsync(c =>
+                        c.CategoryName.ToLower() == predictedCategory.ToLower());
+                    if (category == null)
+                    {
+                        ModelState.AddModelError(string.Empty, "Category could not be predicted. Please enter category");
+                        model.Categories = await _context.Category.ToListAsync();
+                        return View(model);
+                    }
+
+                }
+                else
+                {
+                    category = await _context.Category.FindAsync(model.CategoryId);
+                }
+
+                var news = await _context.News.FindAsync(id);
+                if (news == null) //this shouldn't happen
+                {
+                    return NotFound();
+                }
+
+                news.Category = category;
+                news.PublishedDate = model.PublishedDate;
+                news.Content = model.Content;
+                news.Title = model.Title;
+                news.IsPublished = model.IsPublished;
+                news.FeaturedNews = model.IsFeatured;
+                news.BreakingNews = model.IsBreaking;
+
+                if (model.ImageFile != null)
+                {
+                    if (news.Image != null)
+                    {
+                        var filePath = Path.Combine(webHostEnvironment.WebRootPath, "image/news/", news.Image);
+                        System.IO.File.Delete(filePath);
+                    }
+
+                    news.Image = UploadedFile(model);
+                }
+
+                _context.Update(news);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Details), id);
+            }
+            catch (DbUpdateConcurrencyException ex)
+            {
+                if (!NewsExists(model.Id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    _logger.LogError(ex.Message);
+                }
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex.Message);
+                
+            }
+            return BadRequest("Something Went Wrong!");
         }
 
         // GET: News/Delete/5
@@ -167,7 +285,7 @@ namespace Newsportal.Controllers
                 return Content("No Content Found to delete");
             }
             if (!User.IsInRole("Admin") && news.IsPublished)
-                return Content("cannot delelte published news");
+                return Content("cannot delete published news");
 
             return View(news);
         }
@@ -178,6 +296,16 @@ namespace Newsportal.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var news = await _context.News.FindAsync(id);
+            if (news == null)
+            {
+                return NotFound();
+            }
+            var newsImage = Path.Combine(webHostEnvironment.WebRootPath, "images/news/", news.Image);
+            if (System.IO.File.Exists(newsImage))
+            {
+                System.IO.File.Delete(newsImage);
+            }
+            
             _context.News.Remove(news);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
@@ -188,7 +316,7 @@ namespace Newsportal.Controllers
             return _context.News.Any(e => e.Id == id);
         }
 
-        private string UploadedFile(News model)
+        private string UploadedFile(NewsCreateViewModel model)
         {
             string uniqueFileName = null;
 
@@ -202,7 +330,7 @@ namespace Newsportal.Controllers
                     model.ImageFile.CopyTo(fileStream);
                 }
             }
-            return "/images/" + uniqueFileName;
+            return uniqueFileName;
         }
         
         [HttpPost]
@@ -363,12 +491,10 @@ namespace Newsportal.Controllers
 
             return Ok(rating.Rating);
 
-            /*Random random = new();
-            
-            return Ok(random.Next(1, 6));*/
-
         }
     }
     
-    
+    //implicit operator
+    //        public static implicit operator BalanceType?(string name) => GetInstanceByName<BalanceType>(name);
+
 }
